// lib/interface.ts
export interface simplifiedProduct {
  _id: string;
  imageUrl: string;
  price: number;
  description: string;
  slug: string;
  name: string;
  price_id: string;
}

export interface fullProduct {
  _id: string;
  imageUrl: string;
  price: number;
  slug: string;
  name: string;
  description: string;
  price_id: string;
}